Move some types want to share between sled-agent and nexus.
