+++ sled-agent/src/params.rs
@@ -5,14 +5,15 @@
 use omicron_common::api::internal::nexus::{
     DiskRuntimeState, InstanceRuntimeState,
 };
+use omicron_common::api::internal::shared::{
+    NetworkInterface, SourceNatConfig,
+};
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use std::fmt::{Debug, Display, Formatter, Result as FormatResult};
 use std::net::{IpAddr, Ipv6Addr, SocketAddr, SocketAddrV6};
 use uuid::Uuid;
 
-pub use illumos_utils::opte::params::NetworkInterface;
-pub use illumos_utils::opte::params::SourceNatConfig;
 pub use illumos_utils::opte::params::VpcFirewallRule;
 pub use illumos_utils::opte::params::VpcFirewallRulesEnsureBody;
 pub use sled_hardware::DendriteAsic;
